hal/armv7a: misrify
JIRA: DO-3

+++ hal/armv7a/zynq7000/zynq.h
@@ -19,10 +19,10 @@
 #include "hal/cpu.h"
 
 
-int _zynq_setMIO(unsigned int pin, u8 disableRcvr, u8 pullup, u8 ioType, u8 speed, u8 l0, u8 l1, u8 l2, u8 l3, u8 triEnable);
+int _zynq_setMIO(int pin, u8 disableRcvr, u8 pullup, u8 ioType, u8 speed, u8 l0, u8 l1, u8 l2, u8 l3, u8 triEnable);
 
 
-int _zynq_setAmbaClk(u32 dev, u32 state);
+int _zynq_setAmbaClk(u32 dev, u8 state);
 
 
 void _zynq_interrupts_setCPU(unsigned int irqn, u32 cpuID);
